# api.yaml
...
openapi: 3.1.0
info:
  title: Moodle Replica API
  description: Skeleton API for the people & enrolment module.
  version: 0.1.0
paths:
...
  /api/courses/{course_id}/activities:
    get:
      tags:
      - courses
      summary: List Activities
      description: 'Activities of a course (Team-1 projection reads — bootstrap scope).

        Powers the frontend''s activity-policy table over Mahmoud''s per-activity

        policy endpoint.'
      operationId: list_activities_api_courses__course_id__activities_get
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/enrolment/courses/{course_id}/enrol:
    post:
      tags:
      - enrolment
      summary: Enrol By Course
      description: 'Alias: resolve to a method (given method_id, else the course''s
        manual

        instance) then run the canonical enrol.'
      operationId: enrol_by_course_api_enrolment_courses__course_id__enrol_post
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          title: Course Id
      - name: x-acting-user
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Acting-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseEnrolRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/enrolment/enrolments/{enrolment_id}:
    patch:
      tags:
      - enrolment
      summary: Set Status By Row
      operationId: set_status_by_row_api_enrolment_enrolments__enrolment_id__patch
      parameters:
      - name: enrolment_id
        in: path
        required: true
        schema:
          type: integer
          title: Enrolment Id
      - name: x-acting-user
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Acting-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrolmentStatusPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - enrolment
      summary: Unenrol By Row
      operationId: unenrol_by_row_api_enrolment_enrolments__enrolment_id__delete
      parameters:
      - name: enrolment_id
        in: path
        required: true
        schema:
          type: integer
          title: Enrolment Id
      - name: x-acting-user
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Acting-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CourseEnrolRequest:
      properties:
        user_id:
          type: integer
          title: User Id
        role_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Role Id
        time_start:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Time Start
        time_end:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Time End
        activate:
          type: boolean
          title: Activate
          default: false
        method_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Method Id
      type: object
      required:
      - user_id
      title: CourseEnrolRequest
      description: 'Contract alias POST /courses/{id}/enrol — method optional, defaults
        to

        the course''s manual method.'
...
    EnrolmentStatusPatch:
      properties:
        status:
          type: string
          enum:
          - active
          - suspended
          title: Status
      type: object
      required:
      - status
      title: EnrolmentStatusPatch
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer